+++ puma/plot_base.py
@@ -983,10 +983,14 @@ class PlotBase(PlotObject):
         if labels is None:
             # remove the handles which have label=None
             handles = [handle for handle in handles if handle.get_label() is not None]
+            labels = [handle.get_label() for handle in handles]
+
+        if not handles:
+            return
         ax_mpl.add_artist(
             ax_mpl.legend(
                 handles=handles,
-                labels=([handle.get_label() for handle in handles] if labels is None else labels),
+                labels=labels,
                 loc=self.leg_loc,
                 fontsize=self.leg_fontsize,
                 ncol=self.leg_ncol,