+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### [Latest]
 
+- Enforce xmin and xmax ticks for VarVsVar Plots [#371](https://github.com/umami-hep/puma/pull/371)
+
 ### [v0.5.2](https://github.com/umami-hep/puma/releases/tag/v0.5.2) (02.06.2026)
 
 - Updating Puma Tutorial [#369](https://github.com/umami-hep/puma/pull/369)