+++ changelog.md
@@ -3,6 +3,7 @@
 ### [Latest]
 
 - Enforce xmin and xmax ticks for VarVsVar Plots [#371](https://github.com/umami-hep/puma/pull/371)
+- Update Docker development image [#370](https://github.com/umami-hep/puma/pull/370)
 
 ### [v0.5.2](https://github.com/umami-hep/puma/releases/tag/v0.5.2) (02.06.2026)