+++ changelog.md
@@ -3,6 +3,7 @@
 ### [Latest]
 
 - Adding CLI Flag to Update Reference Plots [#373](https://github.com/umami-hep/puma/pull/373)
+- Allow optional xmin and xmax ticks for PlotBase [#371](https://github.com/umami-hep/puma/pull/371)
 - Fix Issues in with Matplotlib 3.11 [#372](https://github.com/umami-hep/puma/pull/372)
 - Update Docker development image [#370](https://github.com/umami-hep/puma/pull/370)